refactor: warden - core_1 - data passing + typing related improvements

+++ modules/warden/includes/acceleration_readings.h
@@ -4,8 +4,8 @@
 #include "pico_adxl345/adxl345.h"
 
 accelerometer_reading get_accel_readings_mean(accelerometer_reading input[],
-                                               int readings_count);
+                                               size_t readings_count);
 
 void get_bunch_of_accel_readings(ADXL345I2C* adxl345_i2c,
                                  accelerometer_reading total_accel[],
-                                 int readings_count, int delay);
+                                 size_t readings_count, int delay);
